Refactor program
