+++ README.md
@@ -1,4 +1,11 @@
 # NYTSentiment
-A python program for analyzing the sentiment of NYT article headlines.
+A simple Python program for analyzing the sentiment of NYT article headlines.
 
-You'll need to add a valid New York Times API key. It isn't very hard to get one and is free.
+## Running
+You'll need to add a valid New York Times Search API key. One can be acquired easily and for free [here](https://developer.nytimes.com/signup).
+
+Once you've obtained your key, insert it into the environment variable `$NYTKEY` of your shell:
+
+```sh
+export NYTKEY=yourkeyhere
+```
