+++ requirements.txt
@@ -1,0 +1,6 @@
+nltk==3.2.5
+numpy==1.13.3
+pandas==0.21.0
+requests==2.18.4
+scipy==1.0.0
+sklearn==0.0
